fix(web): avoid importing server team.shared in client form
Extract slugifyTeamName to team-slug.ts so team-form.ts does not pull
database dependencies into the client bundle during Next.js build.
