+++ apps/hyperlocalise-web/src/api/routes/team/team-slug.ts
@@ -1,0 +1,7 @@
+export function slugifyTeamName(name: string) {
+  return name
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, "-")
+    .replace(/^-+|-+$/g, "")
+    .slice(0, 120);
+}
